+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 This project is currently at an alpha stage. It will continue on the `0.*.*` branch until the first stable release.  
 
+## develop
+
+### Enhancements:
+
+- `Context\Util\Spin` trait now displays underlying exception message when it times out. 
+
+### Breaking changes:
+- Fix bug in `Context\Util\Spin::fillField()` not calling parent method correctly. This trait can not be used in a `Context` but can be used in page/element objects
+
+
 ## [0.4.0] - 2016-10-21
 
 ### Breaking changes: